many changes

+++ src/main/resources/archetype-resources/src/main/java/__artifactId__/data/DefaultDAO.java
@@ -11,7 +11,7 @@ import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
-import ${package}.${artifactId}.domain.EntityInterface;
+import org.springframework.beans.factory.annotation.Value;
 import ${package}.${artifactId}.domain.IdentifiedEntityInterface;
 
 import java.util.Collections;
@@ -22,6 +22,8 @@ import java.util.List;
  */
 @Repository
 public class DefaultDAO {
+	
+	private @Value("${symbol_dollar}{app.softDeleteEnabled}") boolean softDelete;
 	
 	@Autowired
 	private SessionFactory sessionFactory;
@@ -37,7 +39,7 @@ public class DefaultDAO {
 	 * @param id key to check
      * @return true is entity of the provided class with provided key exists; false otherwise
      */
-	public <ENTITY extends IdentifiedEntityInterface> boolean exists(Class<ENTITY> clazz, Long id) {
+	private <ENTITY extends IdentifiedEntityInterface> boolean exists(Class<ENTITY> clazz, Long id) {
 		
 		if(id == null) {
 			return false;
@@ -46,31 +48,6 @@ public class DefaultDAO {
 		Session session = getSession();
 		
 		return session.get(clazz, id) != null;
-	}
-
-	/**
-	 * Gets an entity by key
-	 *
-	 * @param clazz entity class
-	 * @param id key to search
-	 *
-	 * @return Entity of the provided class with provided ID
-	 *
-	 * @throws ObjectNotExistsException if entity of provided class with provided ID does not exist
-	 * @throws IllegalArgumentException if provided key id NULL
-     */
-	public <ENTITY extends IdentifiedEntityInterface> ENTITY getById(Class<ENTITY> clazz, Long id) throws ObjectNotExistsException {
-		if (id == null) {
-			throw new IllegalArgumentException("Invalid value for entity ID provided:[NULL]");
-		}
-		
-		if(!exists(clazz, id)) {
-			throw new ObjectNotExistsException(clazz.getSimpleName() + " entity with requested id does not exist:[" + String.valueOf(id) + "]");
-		}
-		
-		Session session = getSession();
-
-		return session.load(clazz, id);
 	}
 
 	/**
@@ -99,47 +76,36 @@ public class DefaultDAO {
 	 *
      * @return ID of the newly created entity
      */
-	public <ENTITY extends EntityInterface> Long create(ENTITY entity) {
+	public <ENTITY extends IdentifiedEntityInterface> Long create(ENTITY entity) {
+		if(entity == null) {
+			throw new IllegalArgumentException("Invalid entity provided:[NULL]");
+		}
+		
+		if(entity.getId() != null) {
+			entity.setId(null);
+		}
+		
 		Session session = getSession();
 		
 		return (Long) session.save(entity);
 	}
 
-	/**
-	 * Updates entry information in datastorage by ID of the provided entity.
-	 * @param entity entity information to update datastorage with.
-	 *
-	 * @return true if entity was updated; false otherwise
-	 *
-	 * @throws ObjectNotExistsException if ID of the provided entity is NULL or there is no saved entry
-	 * for entity with ID, equal to ID of the provided entity.
-	 * @throws IllegalArgumentException if provided entity is NULL
-     */
-	public <ENTITY extends IdentifiedEntityInterface> boolean update(ENTITY entity) throws ObjectNotExistsException {
-		if(entity == null) {
-			throw new IllegalArgumentException("Invalid entity provided:[NULL]");
-		}
-		
-		if(entity.getId() == null || !exists(entity.getClass(), entity.getId())) {
-			throw new ObjectNotExistsException(entity.getClass().getSimpleName() + " entity, requested for update does not exist");
-		}
-		
-		Session session = getSession();
-
-		return session.merge(entity) != null;
-	}
-
 	public <ENTITY extends IdentifiedEntityInterface> void delete(Class<ENTITY> clazz, Long id) {
+		if(softDelete) {
+			throw new UnsupportedOperationException("Physical delete operation forbidden");
+		}
+		
 		if (id == null) {
 			throw new IllegalArgumentException("Invalid value for entity ID provided:[NULL]");
 		}
-
+		
 		if(!exists(clazz, id)) {
 			// TODO add logging
-			return;
+			return false;
 		}
-
+		
 		Session session = getSession();
 		session.delete(session.load(clazz, id));
+		return true;
 	}
 }
